Added the tutorial steps as code examples
