+++ examples/tutorial/01-basics.py
@@ -1,0 +1,5 @@
+from agents import *
+
+miner_model = Model("MinerBots", 100, 100)
+
+run(miner_model)
